Changed project structure
Added HollowKnightMP project that bootstraps the core dll. This allows the AppDomain.Current.AssemblyResolve event to take effect where we can load DLL dynamically.
Moved output of HollowKnightMP.Core to HollowKnightMP in a directory HKMP so that deploying the mod is easy by just moving the files over to managed lib.